feat: add support for hardcoded drop down menu + refactor
